Assign panel ids to the samples.

This function assigns panel id's to each sample based on values in the panels and panel_overlap parameters. This is an internal only function.

Usage

PanelDesignAssignPanelids(
  smp,
  panels,
  panel_overlap,
  panel_design,
  number_panels
)

Arguments

smp

A shapefile for the region under study.

panels

A list of integers that define the size of each panel in a non-overlapping panels design. The length of the list determines the number of panels required. The sum of the integers in the panels parameter will determine the total number of samples selected, n. The default value for panels is NULL, this indicates that a non-overlapping panel design is not wanted.

panel_overlap

A list of integers that define the overlap into the previous panel. Is only used when the panels parameter is not NULL. The default value for panel_overlap is NULL. The length of panel_overlap must be equal to the length of panels. The first value is always forced to zero as the first panel never overlaps any region.

panel_design

A flag, when TRUE, indicates that we are performing a panel design and the parameters used are specified in the panels and panel_overlap parameters.

number_panels

The number of sample panels required.

Value

Returns a list of the following variables: sample, which has the appropriate panel id assigned determined by the panels and panel_overlap parameters.

Details

This function was written by Phil Davies.